requirements.txt
async-timeout==4.0.2
awscli==1.22.27
botocore==1.23.27
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
crc16==0.1.1
Deprecated==1.2.13
docutils==0.15.2
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
packaging==21.3
pyasn1==0.4.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-redis==0.3.0
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
